+++ ui/src/kapacitor/constants/index.js
@@ -35,18 +35,6 @@ export const OPERATORS = [
 export const PERIODS = ['1m', '5m', '10m', '30m', '1h', '2h', '24h']
 export const CHANGES = ['change', '% change']
 export const SHIFTS = ['1m', '5m', '10m', '30m', '1h', '2h', '24h']
-export const ALERTS = [
-  'alerta',
-  'hipchat',
-  'opsgenie',
-  'pagerduty',
-  'sensu',
-  'slack',
-  'smtp',
-  'talk',
-  'telegram',
-  'victorops',
-]
 
 export const DEFAULT_RULE_ID = 'DEFAULT_RULE_ID'
 
@@ -82,23 +70,125 @@ export const RULE_MESSAGE_TEMPLATES = {
 
 export const DEFAULT_ALERTS = ['http', 'tcp', 'exec', 'log']
 
-export const DEFAULT_ALERT_LABELS = {
-  http: 'URL:',
-  tcp: 'Address:',
-  exec: 'Add Command (Arguments separated by Spaces):',
-  log: 'File:',
-  smtp: 'Email Addresses (Separated by Spaces):',
-  slack: 'Send alerts to Slack channel:',
-  alerta: 'Paste Alerta TICKscript:',
-}
-export const DEFAULT_ALERT_PLACEHOLDERS = {
-  http: 'Ex: http://example.com/api/alert',
-  tcp: 'Ex: exampleendpoint.com:5678',
-  exec: 'Ex: woogie boogie',
-  log: 'Ex: /tmp/alerts.log',
-  smtp: 'Ex: benedict@domain.com delaney@domain.com susan@domain.com',
-  slack: '#alerts',
-  alerta: 'alerta()',
+export const RULE_ALERT_OPTIONS = {
+  http: {
+    args: {
+      label: 'URL:',
+      placeholder: 'Ex: http://example.com/api/alert',
+    },
+    // properties: [
+    //   {name: 'endpoint', label: 'Endpoint:', placeholder: 'Endpoint'},
+    //   {name: 'header', label: 'Headers:', placeholder: 'Headers (Delimited)'}, // TODO: determine how to delimit
+    // ],
+  },
+  tcp: {
+    args: {
+      label: 'Address:',
+      placeholder: 'Ex: exampleendpoint.com:5678',
+    },
+  },
+  exec: {
+    args: {
+      label: 'Command (Arguments separated by Spaces):',
+      placeholder: 'Ex: woogie boogie',
+    },
+  },
+  log: {
+    args: {
+      label: 'File:',
+      placeholder: 'Ex: /tmp/alerts.log',
+    },
+  },
+  alerta: {
+    args: {
+      label: 'Paste Alerta TICKscript:', // TODO: remove this
+      placeholder: 'alerta()',
+    },
+    // properties: [
+    //   {name: 'token', label: 'Token:', placeholder: 'Token'},
+    //   {name: 'resource', label: 'Resource:', placeholder: 'Resource'},
+    //   {name: 'event', label: 'Event:', placeholder: 'Event'},
+    //   {name: 'environment', label: 'Environment:', placeholder: 'Environment'},
+    //   {name: 'group', label: 'Group:', placeholder: 'Group'},
+    //   {name: 'value', label: 'Value:', placeholder: 'Value'},
+    //   {name: 'origin', label: 'Origin:', placeholder: 'Origin'},
+    //   {name: 'services', label: 'Services:', placeholder: 'Services'}, // TODO: what format?
+    // ],
+  },
+  hipchat: {
+    properties: [
+      {name: 'room', label: 'Room:', placeholder: 'happy_place'},
+      {name: 'token', label: 'Token:', placeholder: 'a_gilded_token'},
+    ],
+  },
+  opsgenie: {
+    // properties: [
+    //   {name: 'recipients', label: 'Recipients:', placeholder: 'happy_place'}, // TODO: what format?
+    //   {name: 'teams', label: 'Teams:', placeholder: 'blue,yellow,maroon'}, // TODO: what format?
+    // ],
+  },
+  pagerduty: {
+    properties: [
+      {
+        name: 'serviceKey',
+        label: 'Service Key:',
+        placeholder: 'one_rad_key',
+      },
+    ],
+  },
+  pushover: {
+    properties: [
+      {
+        name: 'device',
+        label: 'Device:',
+        placeholder: 'dv1,dv2 (Comma Separated)', // TODO: do these need to be parsed before sent?
+      },
+      {name: 'title', label: 'Title:', placeholder: 'Important Message'},
+      {name: 'URL', label: 'URL:', placeholder: 'https://influxdata.com'},
+      {name: 'URLTitle', label: 'URL Title:', placeholder: 'InfluxData'},
+      {name: 'sound', label: 'Sound:', placeholder: 'alien'},
+    ],
+  },
+  sensu: {
+    // TODO: apparently no args or properties, according to kapacitor/ast.go ?
+  },
+  slack: {
+    properties: [
+      {name: 'channel', label: 'Channel:', placeholder: '#cubeoctohedron'},
+      {name: 'iconEmoji', label: 'Emoji:', placeholder: ':cubeapple:'},
+      {name: 'username', label: 'Username:', placeholder: 'pineapple'},
+    ],
+  },
+  smtp: {
+    args: {
+      label: 'Email Addresses (Separated by Spaces):',
+      placeholder:
+        'Ex: benedict@domain.com delaney@domain.com susan@domain.com',
+    },
+    details: {placeholder: 'Email body text goes here'},
+  },
+  talk: {},
+  telegram: {
+    properties: [
+      {name: 'chatId', label: 'Chat ID:', placeholder: 'xxxxxxxxx'},
+      {name: 'parseMode', label: 'Emoji:', placeholder: 'Markdown'},
+      // {
+      //   name: 'disableWebPagePreview',
+      //   label: 'Disable Web Page Preview:',
+      //   placeholder: 'true', // TODO: format to bool
+      // },
+      // {
+      //   name: 'disableNotification',
+      //   label: 'Disable Notification:',
+      //   placeholder: 'false', // TODO: format to bool
+      // },
+    ],
+  },
+  victorops: {
+    properties: [
+      {name: 'routingKey', label: 'Channel:', placeholder: 'team_rocket'},
+    ],
+  },
 }
 
 export const ALERT_NODES_ACCESSORS = {
@@ -107,7 +197,6 @@ export const ALERT_NODES_ACCESSORS = {
   exec: rule => _.get(rule, 'alertNodes[0].args', []).join(' '),
   log: rule => _.get(rule, 'alertNodes[0].args[0]', ''),
   smtp: rule => _.get(rule, 'alertNodes[0].args', []).join(' '),
-  slack: rule => _.get(rule, 'alertNodes[0].properties[0].args', ''),
   alerta: rule =>
     _.get(rule, 'alertNodes[0].properties', [])
       .reduce(