+++ ui/src/shared/components/CustomTimeRangeDropdown.js
@@ -65,7 +65,7 @@ CustomTimeRangeDropdown.propTypes = {
   onApplyTimeRange: func.isRequired,
   timeRange: shape({
     lower: string.isRequired,
-    upper: string.isRequired,
+    upper: string,
   }).isRequired,
 }